requirements.txt
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.3
httptools==0.7.1
httpx==0.24.0
idna==3.11
pillow==12.0.0
pydantic==1.10.24
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.22.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
